added lots of QoL programs
These files are not download and run ready. some setup is required.

instagram scrape gets posts from an account username and outputs json containing the posts' data.

NewSponsorSearch is ran before the other sponsor scripts to get data from the internet to start to be processed by the other files.

sheetsWriter is a template file for integrating google sheets into the scripts.

sponsor 5 pass2 checks the saved URLs from the other scripts and determines if there is a form. this makes it easier to prioritize what to complete first.

sponsor 7 pass3 checks the email address and determines its relevance to sponsorships based on its name. It is not complete for best use yet.

TeamProgress gets the output from the instagram scraper and makes a history document based on the accounts' posts.